requirements.txt
aiohttp~=3.8.1
python-dotenv~=0.20.0
returns~=0.19.0
toolz~=0.11.2